fix: use default options
