+++ core/vibes/soul/components/compare-products/index.tsx
@@ -1,12 +1,13 @@
 'use client';
 
+import clsx from 'clsx';
 import React, { useState } from 'react';
 
 import { Breadcrumb } from '@/vibes/soul/components/breadcrumbs';
 import { CompareDrawer } from '@/vibes/soul/components/compare-drawer';
 import { type Filters } from '@/vibes/soul/components/filter-panel';
 import { Pagination } from '@/vibes/soul/components/pagination';
-import { Product } from '@/vibes/soul/components/product-card';
+import { Product, ProductCardSkeleton } from '@/vibes/soul/components/product-card';
 import { ProductsHeader, type Sort } from '@/vibes/soul/components/products-header';
 import { ProductsList } from '@/vibes/soul/components/products-list';
 
@@ -58,3 +59,22 @@ export const CompareProducts = function CompareProducts({
     </div>
   );
 };
+
+export const CompareProductsSkeleton = () => (
+  <div className="mx-auto max-w-screen-2xl">
+    <div className="relative z-10 pb-10 @container">
+      <div className="px-3 pb-6 pt-24 @xl:px-6 @4xl:pt-32 @5xl:px-20" />
+      {/* Products Header */}
+      <div className="flex flex-wrap items-center justify-between gap-4 bg-background text-foreground">
+        <div className="h-[50px] animate-pulse cursor-pointer rounded-xl" />
+      </div>
+    </div>
+    <div className="w-full bg-background pt-0.5 @container">
+      <div className="mx-auto grid max-w-screen-2xl grid-cols-1 gap-x-5 gap-y-10 px-3 @md:grid-cols-2 @xl:gap-y-10 @xl:px-6 @4xl:grid-cols-3 @5xl:px-20">
+        {Array.from({ length: 5 }).map((_, index) => (
+          <ProductCardSkeleton key={index} />
+        ))}
+      </div>
+    </div>
+  </div>
+);
